Moved RefMint to its own module
