+++ .env.template
@@ -1,0 +1,17 @@
+# You need a websocket from an RPC provider like Infura, Alchemy, or Ankr
+WSS_URL="wss://eth-goerli.g.alchemy.com/v2/<id>"
+
+# You can get these refmint Info from refmint.xyz
+DEVELOPER_API_KEY=""
+PROJECT_ID=""
+CAMPAIGN_ID=""
+
+# Mongo DB URI
+MONGODB_URI="mongodb+srv://<username>:<password>@sandbox.jadwj.mongodb.net/<database>?retryWrites=true&w=majority"
+
+# Set this to a specific block number if you don't want to start from 0. Setting this to 0 may cause an error as 
+# some RPC history only goes back some 2000 blocks on some chains
+INIT_BLOCK=0
+
+# Skip processing the gap between the last block processed and the latest block
+START_FROM_LATEST_BLOCK="false"
