+++ db/connect.mjs
@@ -2,7 +2,7 @@ import { MongoClient } from "mongodb";
 import dotenv from "dotenv";
 dotenv.config();
 
-const connectionString = process.env.ATLAS_URI || "";
+const connectionString = process.env.MONGODB_URI || "";
 
 const client = new MongoClient(connectionString);
