+++ routes/eventlistener.mjs
@@ -2,17 +2,10 @@
 import erc20abi from "../abi/erc20.js";
 import { ethers } from "ethers";
 import express from "express";
-import db from "../db/connect.mjs";
-import { ObjectId } from "mongodb";
-import { loadSettings, saveLastBlock } from "../utils/settings.mjs";
+import { loadSettings, saveLastBlock } from "../modules/settings.mjs";
+import { triggerEvent } from "../modules/refmint.mjs";
 
 var router = express.Router();
-
-// Set RefMint SDK
-// const refmintClient = new refmintSDK.default.Game({
-// 	apiKey: process.env.DEVELOPER_API_KEY,
-// 	baseUrl: refmintSDK.BaseURLOptions.TESTNET
-// });
 
 // The URL should be your Infura or Alchemy Websocket URL
 // example: wss://eth-goerli.g.alchemy.com/v2/<API_KEY>
@@ -31,46 +24,44 @@ var contract = new ethers.Contract(
 // Transfer filter ABI: Transfer(from, to, amount)
 let transferEventFilter = contract.filters.Transfer();
 
+// Filters any Transfer events where "from" address is equal to 0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b
+// let transferEventFilter = contract.filters.Transfer("0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b");
+
+// Filter any Transfer events where "to" address is equal to 0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b
+// let transferEventFilter = contract.filters.Transfer(null, "0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b");
+
+// Load the Settings from Settings from MongoDB database
 let settings = await loadSettings();
 let nextBlockToProcess = settings.lastBlockProcessed + 1;
 
-// If you want to start listening from latest block, ignore this
+// If you want to start listening from latest block, ignore this. Otherwise, 
+// process old / missed blocks from the last processed block stored in the settings
 if (process.env.START_FROM_LATEST_BLOCK === "false") {
+  // Note: Some chains (like Polygon) only have limited history that you can query from RPCs like Infura 
+  // or Alchemy. They only allow you to look at history of the last 1000-5000 blocks. If it's too big, contract.
+  // queryFilter() will fail.
+  let latestBlock = await provider.provider.getBlockNumber();
+  if (latestBlock - 2000 > nextBlockToProcess) {
+    nextBlockToProcess = latestBlock - 2000;
+  }
+
   console.log(`Processing from block ${nextBlockToProcess}`);
   
-  // On start, if there are any things to worry about
   let events = await contract.queryFilter(transferEventFilter, nextBlockToProcess);
   if (events.length > 0) {
     for (let event of events) {
       nextBlockToProcess = event.blockNumber;
+      
+      // Process the event here, however you want to massage the event
   
       let wallet_address = event.args[0];
-      console.log(`Processing Older Transaction From: ${wallet_address}`);
-      
-      // refmintClient.triggerEvent(
-      //   process.env.PROJECT_ID,
-      //   process.env.CAMPAIGN_ID,
-      //   "Transfer", // Custom Event Name
-      //   "TRANSFER", // Custom Event Type
-      //   "ERC20",    // Custom Event Group
-      //   wallet_address
-      // ).then((resp) => {
-      //   //do something...
-      // }).catch(e => {
-      //   console.log(e);
-      // });
+      await triggerEvent(wallet_address);
     }
   
     // Set the last processed block
     await saveLastBlock(settings._id, nextBlockToProcess);
   }
 }
-
-// Filters any Transfer events where "from" address is equal to 0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b
-// let transferEventFilter = contract.filters.Transfer("0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b");
-
-// Filter any Transfer events where "to" address is equal to 0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b
-// let transferEventFilter = contract.filters.Transfer(null, "0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b");
 
 // We start listening for the Transfer Event on the Goerli DAI contract
 contract.on(transferEventFilter, async (args) => {
@@ -81,25 +72,11 @@ contract.on(transferEventFilter, async (args) => {
   //   0n
   // ]
   let wallet_address = args.args[0];
-  console.log(`From: ${wallet_address}`);
 
-  // You can use the RefMint SDK, or call the RefMint Rest API
-  // RefMint SDK (https://docs.refmint.xyz/refmint-sdk/sdk) 
-  // github: https://github.com/ReferralOS/refmint-sdk
-  // REST API endpoints: https://docs.refmint.xyz/refmint-sdk/sdk/endpoints
-  //    - For REST API, x-api-key header required and Developer API Key
-  // refmintClient.triggerEvent(
-  //   process.env.PROJECT_ID,
-  //   process.env.CAMPAIGN_ID,
-  //   "Transfer", // Custom Event Name
-  //   "TRANSFER", // Custom Event Type
-  //   "ERC20",    // Custom Event Group
-  //   wallet_address
-  // ).then((resp) => {
-  //   //do something...
-  // }).catch(e => {
-  //   console.log(e);
-  // });
+  // Process the event here, however you want to massage the event
+  
+  // Forward the event using the RefMint SDK 
+  await triggerEvent(wallet_address);
   
   // Set the last processed block
   settings.lastBlockProcessed = args.log.blockNumber;
